Altera o dto e repository das entidades e cria o service e controllers

+++ src/main/java/br/edu/ifal/sigamais/dto/FrequenciaRequestDTO.java
@@ -1,6 +1,6 @@
 package br.edu.ifal.sigamais.dto;
 
 public record FrequenciaRequestDTO(
-    Long matriculaId,
+    Integer matriculaId,
     Integer faltas
 ) {}
