+++ src/main/java/br/edu/ifal/sigamais/dto/NotaRequestDTO.java
@@ -2,7 +2,7 @@ package br.edu.ifal.sigamais.dto;
 import java.math.BigDecimal;
 
 public record NotaRequestDTO (
-    Long matriculaId,
+    Integer matriculaId,
     BigDecimal valor,
     String tipo
 ) {}
