+++ src/main/java/br/edu/ifal/sigamais/repository/FrequenciaRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import br.edu.ifal.sigamais.model.Frequencia;
 import java.util.List;
 
-public interface FrequenciaRepository extends JpaRepository<Frequencia, Long> {
+public interface FrequenciaRepository extends JpaRepository<Frequencia, Integer> {
 
-    List<Frequencia> findByMatriculaId(Long matriculaId);
+    List<Frequencia> findByMatriculaId(Integer matriculaId);
 }
